Remove source file with a space at the end... Split file with multiple interfaces into multiple files Fix sonar code smells
