--- workers/unity/Packages/io.improbable.gdk.core/Commands/CommandSendStorage .cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 409087e1e2571a348a3410709ac410d4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
