+++ workers/unity/Packages/io.improbable.gdk.core/Commands/ICommandRequestSendStorage.cs
@@ -1,0 +1,10 @@
+using Unity.Entities;
+
+namespace Improbable.Gdk.Core.Commands
+{
+    public interface ICommandRequestSendStorage<in T> : ICommandSendStorage
+        where T : ICommandRequest
+    {
+        void AddRequest(T request, Entity sendingEntity, CommandRequestId requestId);
+    }
+}
