+++ workers/unity/Packages/io.improbable.gdk.core/Commands/ICommandResponseSendStorage.cs
@@ -1,0 +1,8 @@
+namespace Improbable.Gdk.Core.Commands
+{
+    public interface ICommandResponseSendStorage<in T> : ICommandSendStorage
+        where T : ICommandResponse
+    {
+        void AddResponse(T response);
+    }
+}
