+++ workers/unity/Packages/io.improbable.gdk.core/Commands/ICommandSendStorage.cs
@@ -1,0 +1,7 @@
+namespace Improbable.Gdk.Core.Commands
+{
+    public interface ICommandSendStorage
+    {
+        void Clear();
+    }
+}
